requirements.txt
# Core dependencies for narrative transformation system

# LLM APIs
openai>=1.12.0
anthropic>=0.18.0
google-generativeai>=0.3.0

# NLP and text processing
nltk>=3.8.1
vaderSentiment>=3.3.2

# Utilities
python-dotenv>=1.0.0
typing-extensions>=4.9.0

# Data processing
numpy>=1.24.0

# Visualization (optional, for tension curves)
matplotlib>=3.8.0

# Testing (optional)
pytest>=7.4.0
pytest-cov>=4.1.0